+++ soxs/instrument.py
@@ -439,11 +439,13 @@ def make_background(
     diffuse_unresolved : boolean, optional
         Add a diffuse component across the entire field of view to represent
         the unresolved flux from sources at very small fluxes. Default: True
-    drop_brightest : integer, optional
+    drop_brightest : integer or float, optional
         If set to an integer, drop the brightest *drop_brightest* sources
-        from the list. This is a poor-person's way of mimicking the
-        by-hand removal of point sources via wavdetect or some other
-        technique. Default: None
+        from the list. If set to a float, it must be between 0.0 and 1.0, and
+        it is interpreted as the fraction of the total flux that you want dropped
+        in terms of point sources that will determine the number. This is a
+        poor-person's way of mimicking the by-hand removal of point sources via
+        wavdetect or some other technique. Default: None
     aimpt_shift : array-like, optional
         A two-float array-like object which shifts the aimpoint on the
         detector from the nominal position. Units are in arcseconds.
@@ -663,11 +665,13 @@ def make_background_file(
     diffuse_unresolved : boolean, optional
         Add a diffuse component across the entire field of view to represent
         the unresolved flux from sources at very small fluxes. Default: True
-    drop_brightest : integer, optional
+    drop_brightest : integer or float, optional
         If set to an integer, drop the brightest *drop_brightest* sources
-        from the list. This is a poor-person's way of mimicking the
-        by-hand removal of point sources via wavdetect or some other
-        technique. Default: None
+        from the list. If set to a float, it must be between 0.0 and 1.0, and
+        it is interpreted as the fraction of the total flux that you want dropped
+        in terms of point sources that will determine the number. This is a
+        poor-person's way of mimicking the by-hand removal of point sources via
+        wavdetect or some other technique. Default: None
     instr_bkgnd_scale : float, optional
         A constant factor by which to scale the instrumental background
         spectrum up or down. Default: 1, which means no scaling.
@@ -893,11 +897,13 @@ def instrument_simulator(
     diffuse_unresolved : boolean, optional
         Add a diffuse component across the entire field of view to represent
         the unresolved flux from sources at very small fluxes. Default: True
-    drop_brightest : integer, optional
+    drop_brightest : integer or float, optional
         If set to an integer, drop the brightest *drop_brightest* sources
-        from the list. This is a poor-person's way of mimicking the
-        by-hand removal of point sources via wavdetect or some other
-        technique. Default: None
+        from the list. If set to a float, it must be between 0.0 and 1.0, and
+        it is interpreted as the fraction of the total flux that you want dropped
+        in terms of point sources that will determine the number. This is a
+        poor-person's way of mimicking the by-hand removal of point sources via
+        wavdetect or some other technique. Default: None
     instr_bkgnd_scale : float, optional
         A constant factor by which to scale the instrumental background
         spectrum up or down. Default: 1, which means no scaling.